fix: move Score to VO; save `score` through enrollment; `TranscriptRepository` now only support get academic transcript
