+++ backend/src/main/java/com/tkpm/sms/application/mapper/ScoreMapper.java
@@ -3,7 +3,7 @@ package com.tkpm.sms.application.mapper;
 import com.tkpm.sms.application.dto.request.enrollment.TranscriptUpdateRequestDto;
 import com.tkpm.sms.application.dto.response.enrollment.TranscriptDto;
 import com.tkpm.sms.application.dto.response.enrollment.ScoreDto;
-import com.tkpm.sms.domain.model.Score;
+import com.tkpm.sms.domain.valueobject.Score;
 
 public interface ScoreMapper {
     TranscriptDto toScoreDto(Score score);
